// projects/bible/mobile/src/interfaces/bibles/VerseOgTheDay.ts
// Generated by https://quicktype.io

import { IText } from "./FavouritesVerses";

export interface VerseDayResponse {
  ok: boolean;
  code: number;
  status: string;
  data: IVotd;
}

export interface VerseDayResponseList {
  ok: boolean;
  code: number;
  status: string;
  data: IVotd[];
}

export interface IVotd {
  id: number;
  uuid: string;
  version: Version;
  bookName: string;
  bookNumber: number;
  chapter: number;
  verses: string;
  texts: IText[];
  createdAt: string;
  updatedAt: string;
}

export interface Version {
  language: string;
  longLanguage: string;
  name: string;
  description: string;
  file: string;
}